[JBPAPP-9860]:Fix the concurrent issue in tmp xsd file change
git-svn-id: https://anonsvn.jboss.org/repos/jbossws/stack/native/branches/jbossws-native-3.1.2@16781 50344849-fb0f-0410-a53f-a94597aac05c

+++ modules/core/src/main/java/org/jboss/ws/tools/wsdl/WSDL11Reader.java
@@ -267,7 +267,7 @@ public class WSDL11Reader
    {
       for (File current : tempFiles)
       {
-         current.delete();
+         current.deleteOnExit();
       }
    }
 
@@ -580,7 +580,7 @@ public class WSDL11Reader
          }
       }
 
-      if (tmpFile != null)
+      if (tmpFile != null && !tmpFile.exists())
       {
          FileWriter fwrite = new FileWriter(tmpFile);
          new DOMWriter(fwrite).setPrettyprint(true).print(schemaEl);
